requirements.txt
fastapi==0.109.2
fastapi-cache2>=0.2.0
uvicorn==0.27.0
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.28.0
alembic==1.11.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.25.2
redis==4.5.5
python-multipart==0.0.6
bcrypt==4.0.1
pydantic-settings==2.2.1
pydantic==2.6.3
email-validator==2.1.1
aiofiles
minio